Compute BMI using standard formula

simple_bmi Computes BMI using standard formula. Assumes input compatible with output from longwide().

simple_bmi(wide_df, wtcol = "wt", htcol = "ht")

Arguments

wide_df

A data frame or data table containing heights and weights in wide format, e.g., after transformation with longwide()

wtcol

name of observation height value column, default 'wt'

htcol

name of subject weight value column, default 'ht'

Value

Returns a data table with the added column "bmi"

Examples

# Simple usage
# Run on a small subset of given data
df <- as.data.frame(syngrowth)
df <- df[df$subjid %in% unique(df[, "subjid"])[1:5], ]
df <- cbind(df,
            "gcr_result" = cleangrowth(df$subjid,
                                       df$param,
                                       df$agedays,
                                       df$sex,
                                       df$measurement))
# Convert to wide format
wide_df <- longwide(df)
wide_df_with_bmi <- simple_bmi(wide_df)

# Specifying different column names; note that quotes are used
colnames(wide_df)[colnames(wide_df) %in% c("wt", "ht")] <-
  c("weight", "height")
wide_df_with_bmi <- simple_bmi(wide_df, wtcol = "weight", htcol = "height")